Create Auto
How to use: When you click the bookmark/enter the code it will ask you how much puzzles and time you want to set. If you stay on the STMath page without going on another website for example clicking a lesson and then returning to the home page when you click it again it will set the amount you entered.